Rename the createnotes package to notes
