+++ tool/release/notes/BUILD
@@ -24,9 +24,9 @@ load("@io_bazel_rules_kotlin//kotlin:kotlin.bzl", "kt_jvm_binary")
 package(default_visibility = ["//visibility:public"])
 
 kt_jvm_binary(
-    name = "bin",
+    name = "create",
     srcs = glob(["*.kt"]),
-    main_class = "com.vaticle.dependencies.tool.release.createnotes.CreateNotesKt",
+    main_class = "com.vaticle.dependencies.tool.release.notes.NotesCreateKt",
     deps = [
         "@maven//:com_eclipsesource_minimal_json_minimal_json",
         "@maven//:com_google_http_client_google_http_client",
