+++ tool/release/notes/NotesCreate.kt
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package com.vaticle.dependencies.tool.release.createnotes
+package com.vaticle.dependencies.tool.release.notes
 
 import java.nio.file.Path
 import java.nio.file.Paths
@@ -28,7 +28,7 @@ import kotlin.text.Regex.Companion.escapeReplacement
 
 fun main(args: Array<String>) {
     val bazelWorkspaceDir = Paths.get(getEnv("BUILD_WORKSPACE_DIRECTORY"))
-    val githubToken = getEnv("CREATE_NOTES_TOKEN")
+    val githubToken = getEnv("NOTES_CREATE_TOKEN")
     if (args.size != 5) throw RuntimeException("org, repo, commit, version, and template must be supplied")
     val (org, repo, commit, version, templateFileLocation) = args
     val templateFile = bazelWorkspaceDir.resolve(templateFileLocation)
